Make find_file public
